% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackp.R
\name{peek}
\alias{peek}
\title{Looks the top value on the stack}
\usage{
peek(sp)
}
\arguments{
\item{sp}{is the current stack object}
}
\value{
The topmost element in the stack
}
\description{
Looks the top value on the stack
}
\examples{
s <- stackp()
s <- peek()
}
